Extract XML serializer and include it to the Comment model
To reduce the size of our models and reduce clutter, move the xml
serialization to an external module.
